requirements.txt
PyMuPDF==1.24.10
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.40
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.137
openai==1.51.2
python-docx==1.1.2
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
matplotlib==3.9.2
tiktoken
flask
pyyaml
